# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/cart/additem": {
            "post": {
                "description": "添加商品到指定用户的购物车",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "添加商品到购物车",
                "parameters": [
                    {
                        "description": "商品和用户信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.addItemReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/cart/empty": {
            "post": {
                "description": "清空指定用户的购物车",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "清空购物车",
                "parameters": [
                    {
                        "description": "用户id",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.userIdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/cart/get": {
            "get": {
                "description": "获取指定用户的购物车中的商品",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "获取购物车",
                "parameters": [
                    {
                        "description": "用户id",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.userIdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "购物车中的商品",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.cartItem"
                            }
                        }
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/order/cancel": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "parameters": [
                    {
                        "description": "订单id",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.orderIdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/order/list": {
            "get": {
                "description": "获取用户的订单列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "订单列表",
                "parameters": [
                    {
                        "description": "用户id",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.userIdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "订单列表",
                        "schema": {
                            "$ref": "#/definitions/handler.orderListRes"
                        }
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/order/place": {
            "post": {
                "description": "创建一个新的订单",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "创建订单",
                "parameters": [
                    {
                        "description": "订单信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.orderPlaceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "订单id",
                        "schema": {
                            "$ref": "#/definitions/handler.orderIdReq"
                        }
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.addItemReq": {
            "type": "object",
            "required": [
                "item",
                "userid"
            ],
            "properties": {
                "item": {
                    "$ref": "#/definitions/handler.cartItem"
                },
                "userid": {
                    "type": "integer"
                }
            }
        },
        "handler.address": {
            "type": "object",
            "required": [
                "city",
                "country",
                "state",
                "street_address",
                "zip_code"
            ],
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street_address": {
                    "type": "string"
                },
                "zip_code": {
                    "type": "integer"
                }
            }
        },
        "handler.cartItem": {
            "type": "object",
            "required": [
                "product_id",
                "quantity"
            ],
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "handler.errorReturn": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handler.item": {
            "type": "object",
            "required": [
                "cost",
                "product_id",
                "quantity"
            ],
            "properties": {
                "cost": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "handler.orderDetails": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/handler.address"
                },
                "created_at": {
                    "description": "时间戳",
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "user_currency": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "handler.orderIdReq": {
            "type": "object",
            "required": [
                "order_id",
                "user_id"
            ],
            "properties": {
                "order_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "handler.orderListRes": {
            "type": "object",
            "required": [
                "orders"
            ],
            "properties": {
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.orderDetails"
                    }
                }
            }
        },
        "handler.orderPlaceReq": {
            "type": "object",
            "required": [
                "items",
                "user_currency",
                "user_id"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/handler.address"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.item"
                    }
                },
                "user_currency": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "handler.userIdReq": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}